Conflicts:
	PutNavMobile/app/src/main/java/pl/poznan/put/putnav/DatabaseHandler.java

+++ PutNavMobile/app/src/main/java/pl/poznan/put/putnav/DatabaseHandler.java
@@ -2,6 +2,8 @@ package pl.poznan.put.putnav;
 
 import android.content.Context;
 import android.database.sqlite.SQLiteDatabase;
+import android.os.Environment;
+import android.util.Log;
 
 import com.j256.ormlite.android.apptools.OrmLiteSqliteOpenHelper;
 import com.j256.ormlite.dao.Dao;
@@ -9,7 +11,6 @@ import com.j256.ormlite.support.ConnectionSource;
 
 import java.io.File;
 import java.io.FileInputStream;
-import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
@@ -19,6 +20,7 @@ public class DatabaseHandler extends OrmLiteSqliteOpenHelper {
 
     private static final String DATABASE_NAME = "database.db";
     private static int DATABASE_VERSION = 1;
+    private String appDir = new File(Environment.getExternalStorageDirectory().getAbsolutePath(), "putnavadmin").getAbsolutePath();
 
     private Dao<MapPoint, Integer> mapPointDao = null;
     private Dao<MapPointsArcs, Integer> mapPointsArcsDao = null;
@@ -30,35 +32,43 @@ public class DatabaseHandler extends OrmLiteSqliteOpenHelper {
 
     public DatabaseHandler(Context context) throws SQLException {
 
-        super(context, DATABASE_NAME, null, DATABASE_VERSION);
+        super(context, ArchiveFileManager.getDatabasePath(),
+                //new File(Environment.getExternalStorageDirectory().getAbsolutePath(), "putnavadmin/database.db").getAbsolutePath(),
+                null, DATABASE_VERSION);
 
-        File dir = new File(context.getDatabasePath(DATABASE_NAME).getPath());
-        dir.mkdirs(); //żeby nie wywalało błędu ENOENT
+        if(!new File(appDir).exists()) {
+            File dir = new File(context.getDatabasePath(DATABASE_NAME).getPath());
+            dir.mkdirs(); //żeby nie wywalało błędu ENOENT
 
-        File mFile = context.getDatabasePath(DATABASE_NAME);
-        if (mFile.exists()) {
-            mFile.delete();
-        }
+            //File dir = new File(appDir);
+            //dir.mkdir();
 
-        String fileName = context.getDatabasePath(DATABASE_NAME).getPath();
-        InputStream in = null;
-
-        try {
-            in = context.getAssets().open(DATABASE_NAME);
-            FileOutputStream out = null;
-            out = new FileOutputStream(fileName);
-            byte[] buffer = new byte[1024];
-            int length = 0;
-            while ((length = in.read(buffer)) > 0) {
-                out.write(buffer);
+            File mFile = context.getDatabasePath(DATABASE_NAME);
+            if (mFile.exists()) {
+                mFile.delete();
             }
-            in.close();
-            out.flush();
-            out.close();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
 
+            String fileName = context.getDatabasePath(DATABASE_NAME).getPath();
+            InputStream in = null;
+            try {
+                String s = new File(appDir, DATABASE_NAME).getAbsolutePath();
+                //Log.i(DatabaseHandler.class.getName(), s);
+                //in = new FileInputStream(new File(appDir, DATABASE_NAME));
+                in = context.getAssets().open(DATABASE_NAME);
+                FileOutputStream out = null;
+                out = new FileOutputStream(fileName);
+                byte[] buffer = new byte[1024];
+                int length = 0;
+                while ((length = in.read(buffer)) > 0) {
+                    out.write(buffer);
+                }
+                in.close();
+                out.flush();
+                out.close();
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+        }
     }
 
     @Override
